멤버,어드민 시큐리티,권한수정
